REPO-3662: - fixed integration tests for user load driver

+++ server/src/test/java/org/alfresco/bm/integration/test/model/TestRunSummary.java
@@ -19,7 +19,7 @@ public class TestRunSummary
     private Integer resultsSuccess;
     private Integer resultsFail;
     private Integer resultsTotal;
-    private Integer successRate;
+    private Double successRate;
     private List<String> drivers;
 
     public Object get_id()
@@ -172,12 +172,12 @@ public class TestRunSummary
         this.resultsTotal = resultsTotal;
     }
 
-    public Integer getSuccessRate()
+    public Double getSuccessRate()
     {
         return successRate;
     }
 
-    public void setSuccessRate(Integer successRate)
+    public void setSuccessRate(Double successRate)
     {
         this.successRate = successRate;
     }
